Changes in game state

+++ core/src/com/mygdx/game/GameState.kt
@@ -6,16 +6,16 @@ class GameState(
     var doorIsOpen: Boolean = true,
 
     var minionTankFactoryLevel: Int = 1,
-    var minionTankCountInside: Int = 0,
-    var minionTankCountOutside: Int = 0,
+    var minionTankCountInside: Float = 0f,
+    var minionTankCountOutside: Float = 0f,
 
     var minionArcherFactoryLevel: Int = 1,
-    var minionArcherCountInside: Int = 0,
-    var minionArcherCountOutside: Int = 0,
+    var minionArcherCountInside: Float = 0f,
+    var minionArcherCountOutside: Float = 0f,
 
     var minionMinerFactoryLevel: Int = 1,
-    var minionMinerCountInside: Int = 0,
-    var minionMinerCountOutside: Int = 0,
+    var minionMinerCountInside: Float = 0f,
+    var minionMinerCountOutside: Float = 0f,
 
     var resourceTriangle: Int = 100,
     var resourceCircle: Int = 0,
